+++ src/Bugfree.Spo.Analytics.Cli/AssemblyInfo.fs
@@ -6,7 +6,7 @@ open System.Reflection
 [<assembly: AssemblyDescription("")>]
 [<assembly: AssemblyProduct("Bugfree.Spo.Analytics.Cli")>]
 [<assembly: AssemblyCopyright("Copyright (c) 2016-2017 Ronnie Holm <mail@bugfree.dk>")>]
-[<assembly: AssemblyVersion("1.2.1")>]
-[<assembly: AssemblyFileVersion("1.2.1")>]
+[<assembly: AssemblyVersion("1.3.0")>]
+[<assembly: AssemblyFileVersion("1.3.0")>]
 
 do ()